requirements.txt
aniso8601==8.0.0
attrs==19.3.0
boto3==1.11.0
botocore==1.14.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==1.1.15
cryptography==2.8
docutils==0.15.2
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
importlib-metadata==1.3.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.0.2
py-bcrypt==0.4
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.7
python-dateutil==2.6.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
s3transfer==0.3.0
six==1.13.0
swagger-spec-validator==2.4.3
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
